+++ lib/rest-client.js
@@ -1,12 +1,12 @@
 
 const assert = require('assert');
 
-const Request = require('./request-v2');
+const RequestWrapper = require('./util/requestWrapper');
 
 module.exports = class RestClient {
 
   constructor(key, secret, livenet=false, options={}) {
-    this.request = new Request(...arguments);
+    this.request = new RequestWrapper(...arguments);
   }
 
   async placeActiveOrder(params) {
