--- lib/utility.js
@@ -1,9 +1,0 @@
-const {createHmac} = require('crypto');
-
-module.exports = {
-  signMessage(message, secret) {
-    return createHmac('sha256', secret)
-      .update(message)
-      .digest('hex');
-  }
-};
