+++ lib/websocket-client.js
@@ -4,7 +4,7 @@ const WebSocket = require('ws');
 
 const defaultLogger = require('./logger');
 const RestClient = require('./rest-client');
-const { signMessage } = require('./utility');
+const { signMessage } = require('./util/requestUtils');
 
 const wsUrls = {
   livenet: 'wss://stream.bybit.com/realtime',
