+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bybit-api",
-  "version": "1.1.9",
+  "version": "1.2.0",
   "description": "A light node.js wrapper for the Bybit Cryptocurrency Derivative exchange API",
   "main": "index.js",
   "scripts": {
@@ -27,13 +27,9 @@
   "homepage": "https://github.com/tiagosiebler/bybit-api#readme",
   "dependencies": {
     "axios": "^0.20.0",
-    "request": "^2.88.0",
-    "ws": "^7.1.2"
+    "ws": "^7.3.1"
   },
   "devDependencies": {
-    "eslint": "^7.10.0",
-    "eslint-plugin-jest": "^24.0.2",
-    "eslint-plugin-prettier": "^3.1.4",
-    "jest": "^26.4.2"
+    "eslint": "^7.10.0"
   }
 }
